refactor: restructure project into modular package layout
- Moved files into logical subfolders: core/, server/, client/, utils/
- Added __init__.py files to enable package imports
- Updated import statements to use absolute or relative imports
- Improved maintainability and scalability of codebase
