+++ src/core/spell.py
