+++ src/server/server.py
@@ -1,6 +1,6 @@
 import socket
 import pickle
-import utils
+from utils import utils
 
 HOST = '0.0.0.0'
 PORT = 65432
